Added new account header to forest builder page
